feat: update balance data type

+++ apps/core-service/db/migrations/000001_init-transaction-feature.up.sql
@@ -1,8 +1,7 @@
-
 CREATE TABLE IF NOT EXISTS "accounts"  (
     id BIGSERIAL PRIMARY KEY,
     "owner" VARCHAR(255) NOT NULL,
-    "balance" BIGINT NOT NULL,
+    "balance" NUMERIC(20,2) NOT NULL,
     "currency" VARCHAR(3) NOT NULL,
     "created_at" timestamptz  NOT NULL DEFAULT now(),
     "updated_at" timestamptz  NOT NULL DEFAULT now()
@@ -13,7 +12,7 @@ CREATE TABLE IF NOT EXISTS "transfers" (
     id BIGSERIAL PRIMARY KEY,
     from_account_id BIGINT NOT NULL,
     to_account_id BIGINT NOT NULL,
-    amount BIGINT NOT NULL,
+    amount NUMERIC(20,2) NOT NULL,
     created_at timestamptz  NOT NULL DEFAULT now()
 );
 
@@ -22,7 +21,7 @@ CREATE TABLE IF NOT EXISTS "entries" (
     id BIGSERIAL PRIMARY KEY,
     account_id BIGINT NOT NULL,
     type VARCHAR(50) NOT NULL,
-    amount BIGINT NOT NULL,
+    amount NUMERIC(20,2) NOT NULL,
     created_at timestamptz  NOT NULL DEFAULT now()
 );
 
@@ -41,7 +40,6 @@ CREATE INDEX ON "entries" ("account_id");
 CREATE INDEX ON "transfers" ("from_account_id");
 
 CREATE INDEX ON "transfers" ("to_account_id");
-
 
 
 COMMENT ON TABLE "accounts" IS 'Table to store account information';
